Wrapper API for getting device loaction string (#4)
This patch adds a new API to get device location string from DBus by
eid. It queries the LocationCode DBus property to get the value. When
it is not available or empty string, std::nullopt will be returned.

Tested:
Calling this API in examples/scan_endpoints can get the device location
successfully.

Signed-off-by: Jiaqing Zhao <jiaqing.zhao@intel.com>